Created plotting module, updated docs
