+++ docs/api/core.md
@@ -21,11 +21,9 @@
     initialize_scdori_parameters
     load_best_model
     load_scdori_inputs save_model_weights
-    plot_topic_activation_heatmap
     save_regulons
     scDoRI
     train_model_grn
     train_scdori_phases
     trainConfig
-    visualize_downstream_targets
 ```
