+++ docs/api/plotting.md
@@ -1,0 +1,10 @@
+# Plotting
+
+```{eval-rst}
+.. currentmodule:: scdori
+.. autosummary::
+    :toctree: ../generated
+
+    visualize_downstream_targets
+    plot_topic_activation_heatmap
+```
